Исправление парсинга меню + добавление категорий

+++ core/general/models/restaurants.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Text, SmallInteger, ForeignKey, TIMESTAMP, Double
+from sqlalchemy import Text, SmallInteger, ForeignKey, TIMESTAMP, Double, BigInteger
 from sqlalchemy.orm import mapped_column as column, Mapped
 
 from core.general.models.base import Base
@@ -9,7 +9,7 @@ from core.general.models.base import Base
 class Organizations(Base):
     __tablename__ = 'organizations'
 
-    id: Mapped[int] = column(SmallInteger, autoincrement=True, primary_key=True)
+    id: Mapped[int] = column(BigInteger, autoincrement=True, primary_key=True)
     name: Mapped[str] = column(Text, unique=True)
     slug: Mapped[str] = column(Text, unique=True)
     photo: Mapped[str] = column(Text, nullable=False)
@@ -20,7 +20,7 @@ class Restaurants(Base):
 
     id: Mapped[int] = column(SmallInteger, autoincrement=True, primary_key=True)
     slug: Mapped[str] = column(Text, unique=True)
-    organization_id: Mapped[int] = column(SmallInteger, ForeignKey(Organizations.id), nullable=False)
+    organization_id: Mapped[int] = column(BigInteger, ForeignKey(Organizations.id), nullable=False)
     longitude: Mapped[float] = column(Double, nullable=False)
     latitude: Mapped[float] = column(Double, nullable=False)
     rating: Mapped[float] = column(Double)
